+++ app/controllers/fd-listform-constructor.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/controllers/fd-listform-constructor';
