--- app/controllers/fd-visual-listform.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-designer/controllers/fd-visual-listform';
