+++ app/routes/fd-listform-constructor.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/routes/fd-listform-constructor';
