+++ app/templates/fd-listform-constructor.hbs
@@ -1,4 +1,4 @@
-<h3 class="ui header">{{t 'forms.fd-visual-listform.caption'}}</h3>
+<h3 class="ui header">{{t "forms.fd-listform-constructor.caption"}}</h3>
 {{#if errorMessages}}
   {{#each errorMessages as |currentError|}}
     {{ui-message
@@ -25,7 +25,7 @@
         {{/unless}}
       {{/unless}}
       <button class="ui button">
-        {{t "forms.fd-visual-listform.preview-mobile-variant"}}
+        {{t "forms.fd-listform-constructor.preview-mobile-variant"}}
       </button>
       <button type="submit" class="ui button close-button" {{action 'close'}}>
         {{t "forms.edit-form.close-button-text"}}
@@ -65,11 +65,11 @@
         <div class="field">
           {{
             flexberry-field
-            placeholder=(t 'forms.fd-visual-listform.name')
+            placeholder=(t "forms.fd-listform-constructor.name")
             readonly=readonly
             required=true
             value=formClass.name
-            label=(t 'forms.fd-visual-listform.name')
+            label=(t "forms.fd-listform-constructor.name")
           }}
         </div>
